Refactor the cpcode matcher
The old one didn't work at all, it only looked at the `X-Cache-Key`
header, rather than decoding the values within it.

This change splits out all the values from that header, then validates
that the passed-in cpcode is present in the cpcode field in that header.

It also uses a custom failure message to return the actual cpcode found
to the caller, using thread local storage.

Finally, it removes the 200 response code requirement, as the test is
about cpcodes, not response status.