Clean out "internal" Akamai functionality headers
Akamai uses a bunch of functionality behind the scenes which they mix
into their product that is never passed onto the downstream. A few of
these things are:

- Ensuring revalidation with the origin (`must-revalidate`)
- Handle serving cached copies of resources should the origin be
  unavailable (`stale-if-error`)
- Refreshing content asynchronously after expiry
  (`stale-while-revalidate`)

The issue with this is that when you apply these cache control values to
your origin and the traffic is coming via another provider, the
`honour_origin_headers` matcher will fail due to Akamai swallowing up
these values in their products but other providers let them pass.

This updates the functionality to drop these cache control values so
that `akamai-rspec` doesn't cry when they are mismatching.